Redo stack views with basic set up
Here, top and bottom stack views are created with no content and no constraints. They are then embedded in a vertical stack view, which has centre x and centre y constraints. Each stack view has centre alignment, equal spacing, and spacing of 10 (vertical) or 8 (horizontal). This layout is working to centre the addition of labels to both the top and bottom horizontal stacks.

The next step is to start doing this programmatically for an array.count, starting with the styling of a label programmatically.