Add a third stack view and add code for text to jump lines
The if/else for jumping lines was previously faulty. It judged whether the line width could accommodate the label width, if not, that label went to the next line. But when the next comes along, it could pass that width test if it is a shorter width. This meant words could be appearing out of order.